#33 Handle get requests & give available timeslots
Made all local entities use Long instead of Integer. As it is in the DB.
Updated & tested bookingCreateRequest to work with Frontend.
Implemented changes from #34 into booking service.
Changed so ResponseHandler can handle JSON replies.

+++ Middleware/bookings/src/main/java/com/toothtrek/bookings/repository/OfficeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.toothtrek.bookings.entity.Office;
 
-public interface OfficeRepository extends JpaRepository<Office, Integer> {
+public interface OfficeRepository extends JpaRepository<Office, Long> {
 
 }
